moved token URL fetch to shared lib
