+++ packages/core/utils/oauth.ts
@@ -1,5 +1,5 @@
 import type { Integration } from '@superglue/client';
-import { getOAuthConfig, integrations, type IntegrationConfig } from '@superglue/shared';
+import { getOAuthConfig, getOAuthTokenUrl, integrations, type IntegrationConfig } from '@superglue/shared';
 import { logMessage } from './logs.js';
 
 export interface OAuthTokens {
@@ -32,28 +32,6 @@ export function isTokenExpired(integration: Integration): boolean {
 }
 
 /**
- * Get token URL for an integration based on its configuration
- */
-export function getTokenUrl(integration: Integration): string | null {
-    // First priority: User-provided token URL in credentials
-    const customTokenUrl = integration.credentials?.token_url as string;
-    if (customTokenUrl) return customTokenUrl;
-
-    // Second priority: Known integration template token URL
-    const match = Object.entries(integrations).find(([key]) =>
-        integration.id === key || integration.urlHost?.includes(key)
-    );
-
-    if (match) {
-        const [_, config] = match as [string, IntegrationConfig];
-        return config.oauth?.tokenUrl || null;
-    }
-
-    // Fallback: Default OAuth token endpoint
-    return `${integration.urlHost}/oauth/token`;
-}
-
-/**
  * Refresh OAuth tokens for an integration
  */
 export async function refreshOAuthToken(
@@ -72,7 +50,7 @@ export async function refreshOAuthToken(
     }
 
     try {
-        const tokenUrl = getTokenUrl(integration);
+        const tokenUrl = getOAuthTokenUrl(integration);
         if (!tokenUrl) {
             throw new Error('Could not determine token URL for integration');
         }
@@ -172,7 +150,7 @@ export async function handleOAuthCallback(
         }
 
         // Get token URL
-        const tokenUrl = getTokenUrl(integration);
+        const tokenUrl = getOAuthTokenUrl(integration);
         if (!tokenUrl) {
             return {
                 success: false,
